+++ src/graysvr/CObjBase.h
@@ -308,8 +308,8 @@ public:
 	// Location
 	virtual bool MoveTo(CPointMap pt, bool bForceFix = false) = 0;	// Move to a location at top level.
 
-	virtual bool MoveNear( CPointMap pt, int iSteps = 0 );
-	virtual bool MoveNearObj( const CObjBaseTemplate *pObj, int iSteps = 0 );
+	virtual bool MoveNear( CPointMap pt, WORD iSteps = 0 );
+	virtual bool MoveNearObj( const CObjBaseTemplate *pObj, WORD iSteps = 0 );
 
 	void inline SetNamePool_Fail( TCHAR * ppTitles );
 	bool SetNamePool( LPCTSTR pszName );
@@ -1098,7 +1098,7 @@ public:
 		return MoveToUpdate( pt, bForceFix);
 	}
 	bool MoveToCheck( const CPointMap & pt, CChar * pCharMover = NULL );
-	virtual bool MoveNearObj( const CObjBaseTemplate *pItem, int iSteps = 0 );
+	virtual bool MoveNearObj( const CObjBaseTemplate *pItem, WORD iSteps = 0 );
 
 	CItem* GetNext() const
 	{
@@ -3216,11 +3216,11 @@ public:
 		FixClimbHeight();
 	}
 	bool MoveToValidSpot(DIR_TYPE dir, int iDist, int iDistStart = 1, bool bFromShip = false);
-	virtual bool MoveNearObj( const CObjBaseTemplate *pObj, int iSteps = 0 )
+	virtual bool MoveNearObj( const CObjBaseTemplate *pObj, WORD iSteps = 0 )
 	{
 		return CObjBase::MoveNearObj(pObj, iSteps);
 	}
-	bool MoveNear( CPointMap pt, int iSteps = 0 )
+	bool MoveNear( CPointMap pt, WORD iSteps = 0 )
 	{
 		return CObjBase::MoveNear(pt, iSteps);
 	}
